feat: major performance optimization and multi-namespace Temporal architecture
Performance - Parallel INSERT Operations:
- Convert serial INSERTs to parallel using worker pools in indexer activities:
  - validators.go: 4 parallel inserts (validators, non-signing, committee_validators, double-signing)
  - pools.go: 2 parallel inserts (pools, pool_holders)
  - dex_batch.go: 3 parallel inserts (orders, deposits, withdrawals)
  - committees.go: 2 parallel inserts (committees, payments)
- Increase connectionsPerActivity from 2 to 8 to support parallel operations
- Update connection pool formula: (parallel_blocks × 20 activities × 8 conns) + 100 buffer

ClickHouse Server Optimization:
- Scale max_connections 50k → 250k and max_concurrent_queries 30k → 100k for 100+ chains
- Enable async_insert with wait_for_async_insert=1 for data consistency
- Configure memory-efficient aggregation with disk spill thresholds (4GB)
- Tune background pool sizes and merge settings for development profile
- Remove insert_quorum (not needed with async_insert wait mode)

Schema Optimization - Nullable Column Removal:
- Replace Nullable columns with DEFAULT values in tx.go and event.go models
- Eliminates UInt8 null-mask overhead per column (ClickHouse optimization)
- Downsize integer types where appropriate (UInt64 → UInt16 for chain_id, committee_id)

Temporal Multi-Namespace Architecture:
- Add TemporalManager for centralized client lifecycle management
- Extract AdminClient and ChainClient into separate files
- Support per-chain namespaces (chain_<id>) with simplified queue names
- Migrate admin app to use TemporalManager

Staging Cleanup Redesign:
- Replace per-block cleanup workflows with hourly batch cleanup
- New CleanStagingBatch method: single DELETE per table instead of one per height
- Reduces mutation count from ~20 per block to ~20 per hour
- Add dedicated MaintenanceWorker to avoid blocking ops queue

Workflow Improvements:
- Fix local activity timeouts: use StartToCloseTimeout instead of ScheduleToCloseTimeout
- Increase HeadScan/GapScan timeouts for reliability
- Use ChainClient references instead of TemporalClient method calls

New Admin Features:
- Add delete chain workflow and activity for chain removal
- Extend admin API with chain management endpoints
- Remove redundant InitializeDB calls from NewChainDb (handled at startup)

Infrastructure Updates:
- Update Tiltfile for improved local development
- Update deployment.yaml with correct connection pool formula comments
- Configure admin-web-proxy for new endpoints
- Adjust PVC sizes for local ClickHouse development

+++ app/indexer/activity/params.go
@@ -133,7 +133,7 @@ func convertRpcParamsToEntity(rpcParams *fsm.Params, height uint64, blockTime ti
 	if rpcParams.Consensus != nil {
 		params.BlockSize = rpcParams.Consensus.BlockSize
 		params.ProtocolVersion = rpcParams.Consensus.ProtocolVersion
-		params.RootChainID = rpcParams.Consensus.RootChainId
+		params.RootChainID = uint16(rpcParams.Consensus.RootChainId)
 		params.Retired = rpcParams.Consensus.Retired
 	}
